separate metrics into subpkg
