+++ src/runtime/metrics/meter.js
@@ -1,0 +1,24 @@
+
+'use strict'
+
+const { createMetric, metricTypes } = require('./utils')
+
+/** @typedef {import('./types').Meter} Meter */
+
+/**
+ * @param {string} name
+ * @returns {Meter}
+ */
+function newMeter (name) {
+  return {
+    toMetric: () => {
+      const metric = createMetric(name, metricTypes.meter, {})
+      // TODO
+      return metric
+    }
+  }
+}
+
+module.exports = {
+  newMeter
+}
