+++ src/runtime/types.ts
@@ -1,5 +1,6 @@
 import type { IPv4, IPv6 } from 'ipaddr.js'
 import type { Logger } from 'winston'
+import type { Metrics } from './metrics'
 
 export interface RunParams {
   testBranch: string
@@ -37,65 +38,9 @@ export interface Events {
   recordCrash: (err: Error) => void
 }
 
-export interface Metric {
-  ts: number
-  type: number
-  name: string
-  measures: Record<string, any>
-}
-export interface MetricsApi {
-  recordPoint: (name: string, value: number) => void
-  counter: (name: string) => Counter
-  ewma: (name: string, alpha: number) => EWMA
-  meter: (name: string) => Meter
-  timer: (name: string) => Timer
-}
-
-export interface Metrics {
-  d: MetricsApi
-  r: MetricsApi
-  close: () => void
-}
-
 export interface RunEnv extends Events, RunParams, Metrics {
   logger: Logger
   runParams: RunParams
   getSignalEmitter: () => SignalEmitter|null
   setSignalEmitter: (e: SignalEmitter) => void
-}
-
-export interface toMetric {
-  toMetric: () => Metric
-}
-export interface Point extends toMetric {
-  value: () => number
-}
-
-export interface Counter extends toMetric {
-  clear: () => void
-  count: () => number
-  dec: (i: number) => void
-  inc: (i: number) => void
-}
-
-export interface EWMA extends toMetric {
-  rate: () => number
-  tick: () => void
-  update: (n: number) => void
-}
-
-export interface Gauge extends toMetric {
-  update: (n: number) => void
-  value: () => number
-}
-
-export interface Histogram extends toMetric {
-
-}
-export interface Meter extends toMetric {
-
-}
-
-export interface Timer extends toMetric {
-
 }
